splitting model unit tests
